MasterVoter/requirements.txt
getmac >= 0.8
mysql-connector-python >=